improvements

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: fde-capu <fde-capu@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/02 16:19:33 by fde-capu          #+#    #+#             */
-/*   Updated: 2020/08/20 17:48:48 by fde-capu         ###   ########.fr       */
+/*   Updated: 2020/08/21 10:56:02 by fde-capu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -289,7 +289,7 @@ t_mat			*matrixx(t_mat *old, t_mat *new);
 char			*primitive_type_number(int type);
 t_ray			*ray_build(t_vec *a, t_vec *b);
 void			ray_destroy(t_ray *ray);
-int				hit_sphere(t_ray *ray, t_prm *sphere);
+double			hit_sphere(t_ray *ray, t_prm *sphere);
 
 /*
 ** Debugs:
